+++ src/presentations/components/LoginPage/components/Footer/index.jsx
@@ -1,0 +1,10 @@
+export const Footer = () => {
+  return (
+    <footer
+      className="position-absolute bottom-0 text-center w-100 py-2"
+      style={{ backgroundColor: "#f8f9fa" }}
+    >
+      <p className="mb-0 fs-6 fw-light">&copy; 2024 Tsubasa Watanabe</p>
+    </footer>
+  );
+};
