+++ src/presentations/components/LoginPage/components/LoginForm/hooks/useLogin.js
@@ -1,0 +1,32 @@
+import { useState } from "react";
+import { supabase } from "../../../../../../utils/supabase";
+
+export const useLogin = () => {
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
+  const [error, setError] = useState("");
+
+  const handleLogin = async (e) => {
+    e.preventDefault(); // デフォルトの挙動をキャンセル
+
+    const { error } = await supabase.auth.signInWithPassword({
+      email,
+      password,
+    });
+
+    if (error) {
+      setError("ログインに失敗しました");
+    } else {
+      setError("");
+    }
+  };
+
+  return {
+    email,
+    setEmail,
+    password,
+    setPassword,
+    error,
+    handleLogin,
+  };
+};
