+++ src/presentations/components/LoginPage/components/index.js
@@ -1,0 +1,2 @@
+export * from "./Footer";
+export * from "./LoginForm";
